+++ src/protocols/SinglePixel.cpp
@@ -70,10 +70,6 @@ CSinglePixelBufferResource::CSinglePixelBufferResource(uint32_t id, wl_client* c
     });
 }
 
-CSinglePixelBufferResource::~CSinglePixelBufferResource() {
-    PROTO::singlePixel->destroyResource(this);
-}
-
 bool CSinglePixelBufferResource::good() {
     return buffer->good();
 }
