+++ src/protocols/core/Compositor.hpp
@@ -16,6 +16,7 @@
 #include "../../helpers/math/Math.hpp"
 #include "../types/Buffer.hpp"
 #include "../types/SurfaceRole.hpp"
+#include "../types/SurfaceState.hpp"
 
 class CWLOutputResource;
 class CMonitor;
@@ -51,34 +52,6 @@ class CWLRegionResource {
 
   private:
     SP<CWlRegion> resource;
-};
-
-struct SSurfaceState {
-    CRegion                opaque, input = CBox{{}, {INT32_MAX, INT32_MAX}}, damage, bufferDamage = CBox{{}, {INT32_MAX, INT32_MAX}} /* initial damage */;
-    wl_output_transform    transform = WL_OUTPUT_TRANSFORM_NORMAL;
-    int                    scale     = 1;
-    SP<CHLBufferReference> buffer; // buffer ref will be released once the buffer is no longer locked. For checking if a buffer is attached to this state, check texture.
-    SP<CTexture>           texture;
-    Vector2D               offset;
-    Vector2D               size, bufferSize;
-    struct {
-        bool     hasDestination = false;
-        bool     hasSource      = false;
-        Vector2D destination;
-        CBox     source;
-    } viewport;
-    bool rejected  = false;
-    bool newBuffer = false;
-
-    //
-    void reset() {
-        damage.clear();
-        bufferDamage.clear();
-        transform = WL_OUTPUT_TRANSFORM_NORMAL;
-        scale     = 1;
-        offset    = {};
-        size      = {};
-    }
 };
 
 class CWLSurfaceResource {
